requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
PyYAML>=6.0.0